Added color functionality and some other stuff
In the board file, there is now an extra column for all properties
describing the "color" of the space.  This isn't stored as the actual
color value, but rather as simply a character.  It will be used to test
for monopolies, which have an effect on things like rent and whether or
not houses can be built.

I also added a bunch more stuff, mostly get/set functions, but they
aren't fully working yet.

KNOWN ERRORS:
Derived classes not overriding base class for functions, not sure why.
I'm still very new to the whole inheritance thing, so I'll have to look
into it at a later date.  This is most likely my last commit of the day,
unless I get really bored tonight or can't sleep.