Card definition, distribution calc rewrite, more simulation work
Chance and Community Chest cards are now defined in their own files and
read in by monopoly.py.  Each card has a given "card type" denoted by a
2-character string which represents the type of action that the card
causes the drawing player to do

distribution.cpp has been revisited and rewrote to take into account the
change in frequency of jail visits due to rolling doubles 3 times in a
row (important because of impact on amount of time spent on other board
spaces), and the change in the amount of time spent in jail (not
necessarily important because it doesn't change the amount of time spent
on owned spaces relative to each other, but I think it may be useful in
the future)

Work on monopoly.py has been continued as well, with new classes such as
Card and Container, but it is still not at the point of being
functional.